+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 edbee.lib:
 
+- fix #67, PlacholderText support via TextEditorWidget::setPlaceholderText. (uses 70% opacity of foreground color)
 - fix #98, Missing header include in Qt 5.15rc
 - fix #94 (partial), Resource delete fix in TextDocumentscopes
 - Merged PR #86, Updated QsLog